LAUNCH
So we managed to setup the raspberry pi and run it. The problem was most propably with the python version listed.
I reverted back to using "setup-python@v4" from v3 and switched python version from "3.12.0" to "3.12". Bot works without problems right now but there is auto-reboot functionality missing which is another days problem.